@@ -1,11 +1,12 @@
 package com.example.uhf.fragment;
 
-import android.support.v4.app.Fragment;
+import androidx.fragment.app.Fragment;
 
 /**
  * Created by Administrator on 2015-03-10.
  */
-public class KeyDwonFragment extends Fragment {
+public class KeyDwonFragment extends Fragment
+{
 
     public void myOnKeyDwon() {
 
