@@ -2,7 +2,7 @@ package com.U8.reader.server;
 
 import android.content.Context;
 import android.content.Intent;
-import android.support.v4.content.LocalBroadcastManager;
+import androidx.core.content.LocalBroadcastManager;
 import android.util.Log;
 
 import com.U8.operation.U8Series;
